+++ jdaviz/configs/default/plugins/export/export.vue
@@ -33,7 +33,7 @@
           :items="viewer_format_items.map(i => i.label)"
           label="Format"
           hint="Image format for exporting viewers."
-          :disabled="viewer_selected.length == 0" 
+          :disabled="viewer_selected.length == 0"
           persistent-hint
         >
         </v-select>
@@ -126,9 +126,9 @@
     </div>
 
     <div v-if="subset_items.length > 0">
-      <span class="export-category">Spatial Subsets</span>
+      <span class="export-category">Subsets</span>
       <v-row>
-        <span class="category-content v-messages v-messages__message text--secondary">Export spatial subset as astropy region.</span>
+        <span class="category-content v-messages v-messages__message text--secondary">Export spatial subset as astropy region or spectral subset as specutils SpectralRegion.</span>
       </v-row>
       <plugin-inline-select
         :items="subset_items"
@@ -150,7 +150,9 @@
           :menu-props="{ left: true }"
           attach
           v-model="subset_format_selected"
-          :items="subset_format_items.map(i => i.label)"
+          :items="subset_format_items"
+          item-text="label"
+          item-disabled="disabled"
           label="Format"
           hint="Format for exporting subsets."
           :disabled="subset_selected == null || subset_selected.length == 0"
@@ -200,10 +202,10 @@
         :single_select_allow_blank="false"
       >
       </plugin-inline-select>
-      <jupyter-widget 
+      <jupyter-widget
           v-if='plugin_plot_selected_widget.length > 0'
           style="position: absolute; left: -100%"
-          :widget="plugin_plot_selected_widget"/> 
+          :widget="plugin_plot_selected_widget"/>
       <v-row class="row-min-bottom-padding">
         <v-select
           class="category-content"
@@ -222,7 +224,7 @@
 
     <div style="display: grid; position: relative"> <!-- overlay container -->
     <div style="grid-area: 1/1">
-      
+
     <plugin-auto-label
       :value.sync="filename_value"
       :default="filename_default"
@@ -297,7 +299,7 @@
    display: block;
    text-align: center;
    overflow: hidden;
-   white-space: nowrap; 
+   white-space: nowrap;
    text-transform: uppercase;
    color: gray;
    font-weight: 500;